requirements.txt
pandas==2.1.4
mysql-connector-python==8.2.0
python-dotenv==1.0.0
apache-airflow==2.8.0
apache-airflow-providers-mysql==3.4.0
sqlalchemy==2.0.23
numpy==1.26.0
python-dateutil==2.8.2
tqdm==4.66.1
streamlit==1.28.0
plotly==5.17.0
